IndGOterm {IndGOterm}	R Documentation

The main function of individPath

Description

The function is used to compute individual dyregulated pathway.

Usage


IndGOterm(
  stable.pair,
  term,
  patients,
  threads = 1L,
  gene.minimum = 1L,
  threshold = 0.05,
  term_gene_list,
  non_redun
)

Arguments

stable.pair
a matrix or data.frame of stable pairs from normal samples with two columns, the first column is the higher expresion gene , the second column is the lower expresion gene.By default the variables are taken from environment(formula).

patients
a dataset containing tumor samples , the first column must be the geneID ,and tumor samples start with the second column.By default the variables are taken from environment(formula).

threads
an integer value to make sure how many threads you will use to complete the computation

threshold
a numeric value which is used to control false discovery rate under the p_value of the chip-square test or fisher exact probability test , default is 0.05.

non_redun

Value

a list contained two data.frame , individual dyregulated pathway with term ID and individual dyregulated pathway with term name

[Package IndGOterm version 1.0.0 Index]