get.globdysterm {IndGOterm}	R Documentation

get holistic dyregulated pathway

Description

the function is used to get global dyregulated terms from a host of disease samples this function support parallel computation . You need to set thread numbers to make sure how many threads do you need to use .

Usage


get.globdysterm(
  stable.pair,
  reverse.pair,
  threads = 1L,
  gene.minimum = 1,
  term_gene_list = BPtermgene_list,
  patients = tumor_sample,
  threshold = 0.05
)

Arguments

stable.pair
a matrix or data.frame of stable pairs from normal samples with two columns, the first column is the higher expresion gene , the second column is the lower expresion gene.By default the variables are taken from environment(formula).

reverse.pair
a matrix or data.frame of reverse pairs from disease samples with two columns, the first column is the higher expresion gene , the second column is the lower expresion gene.By default the variables are taken from environment(formula).

threads
an integer value to make sure how many threads you will use to complete the computation

patients
a dataset containing tumor samples , the first column must be the geneID ,and tumor samples start with the second column.By default the variables are a dataset with 2003 colorectal cancer samples taken from environment(formula).

threshold
a numeric value which is used to control false discovery rate under the p_value of the chip-square test or fisher exact probability test , default is 0.05.

term.gene
a character string specifying the category of term gene .It can only be one of "BPtermgene","CCtermgene" or "MFtermgene"

Value

a data.frame containing three columns represent "termID" , "p value" , "fdr value" respectively

[Package IndGOterm version 1.0.0 Index]