remove.redundance {IndGOterm}	R Documentation

remove redundant pathway

Description

this function is used to remove redundant parent or son pathway from all of global dyregulated terms

Usage


remove.redundance(
  stable.pair,
  reverse.pair,
  gene.minimum = 1,
  patients,
  re_graph,
  holidyregpath,
  term_gene_list,
  threads = 1L
)

Arguments

stable.pair
a matrix or data.frame of stable pairs from normal samples with two columns, the first column is the higher expresion gene , the second column is the lower expresion gene.By default the variables are taken from environment(formula).

reverse.pair
a matrix or data.frame of reverse pairs from disease samples with two columns, the first column is the higher expresion gene , the second column is the lower expresion gene.By default the variables are taken from environment(formula).

patients
a dataset containing tumor samples , the first column must be the geneID ,and tumor samples start with the second column.By default the variables are a dataset with 2003 colorectal cancer samples taken from environment(formula).

re_graph
a matrix or data.frame with three columns about the relationship between parent term and whose son term. The first to third column is parent termID, son termID and distance between parent term and son term separately

holidyregpath
a matrix or data.frame with three columns depicts the coorelation levels of terms and disease. Which computed by method get.holidyregpath. The first to third column is termID, p-value and FDR value.

term_gene_list
a list containing term genes. each list name is termID and the content of each list is the genes containing in the term.

threads
an integer value to make sure how many threads would be used to complete the computation

Value

a list.The first list is all terms and their redundant parent terms. of all the nonredundant holistic-dyregulated pathway

[Package IndGOterm version 1.0.0 Index]