revpairs {IndGOterm}	R Documentation

get reverse pairs from all the tumor samples

Description

the function is used to get reverse pairs from a host of disease samples. this function support parallel computation . You need to set thread numbers to make sure how many threads do you need to use .

Usage


revpairs(
  stable.pair,
  patients,
  threshold = 0.05,
  spairs_threshold = 0.99,
  threads = 1L,
  capacity = 300000L
)

Arguments

stable.pair
a matrix or data.frame of stable pairs from normal samples with two columns, the first column is the higher expresion gene , the second column is the lower expresion gene.By default the variables are taken from environment(formula).

patients
a dataset containing tumor samples , the first column must be the geneID ,and tumor samples start with the second column.By default the variables are taken from environment(formula).

threshold
a numeric value which is used to control false discovery rate under the p_value of the chip-square test or fisher exact probability test , default is 0.05.

spairs_threshold
a threshold used in function "spairs" to stipulate the circumstance of retaining a gene pair as a stable pair

threads
an integer value to make sure how many threads you will use to complete the computation

capacity
an integer value to depict the computation capacity, ruling how many lines of stable pairs would be computed withiin one time.the default is 300000

Value

a matrix containing four columns respectively represent higher expression gene , lower expression gene , p_value under binomial distribution , false discovery rate under p.adjust

[Package IndGOterm version 1.0.0 Index]