spairs {IndGOterm}	R Documentation

get stable pairs from normal samples

Description

the function is used to get stable pairs from a host of normal samples.

Usage


spairs(data1, gid1, threshold = 0.99)

Arguments

data1
a (normal sample) dataset which is used to screen the stable_pair.Notable that rownames is gene identifier(geneID/genesymbol/ensembleID).

gid1
a vector of gene indentifier which can completely match data1

threshold
a threshold value is used to stipulate that a gene pair maintain the circumstance of > the threshold so that the gene pair should be retained.The threshold should between 0-1, and the default is 0.99.

Value

a list which contains a dataframe of stable pair and a runtime

[Package IndGOterm version 1.0.0 Index]